streamlit.app/requirements.txt
streamlit==1.45.1
Pillow==11.1.0
torch==2.10.0
torchvision==0.25.0
